Add single click functionality
Often times when using calltree I want to see where the call is being used in the current function or any of the child functions. This allows for doing so. Single click will show you the function without updating the calltree. Double click will continue to work as usual.

Other changes made:
1. Switched from `notifyOffsetChanged` to `notifyViewLocationChanged` because `notifyOffsetChanged` is called multiple times on a single/double click and it was making the calltree update when not desired. Also Binary Ninja recommends using this method due to granularity on what changed.
2. Switched `prev_func_offset` to `prev_location` to track new location data.
3. Updated `update_widget` to set the cur_func that is being passed instead of setting it, then updating it passing it the same function.
4. Made `BNFuncItem` not be editable since it should not be.
5. Removed duplicate `func_depth` property